// src/app/api/model/post.ts
/**
 * Sports fever
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: Shankarmurthy.bhat@in.bosch.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Post {
    id?: string;
    userId?: string;
    dateTime?: string;
    text?: string;
    imageId?: string;
}